requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
pywhatkit>=5.3
scikit-image>=0.18.0
scikit-learn>=0.24.0
requests>=2.25.0
tqdm>=4.50.0 